requirements.txt
asgiref==3.8.1
coverage==7.6.4
Django==4.2.16
django-fsm==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
gitdb==4.0.11
GitPython==3.1.43
iniconfig==2.0.0
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
